% This file was not automatically generated by Roxygen2.
% It was written on GitHub's file editor on a phone.

\name{tracestack}
\alias{tracestack}
\title{Search Stack Overflow for your most recent error message.}
\usage{tracestack()}
\description{Search Stack Overflow for your most recent error message.}
